vulcan FE - dashboard code reorganization
