+++ vulcan/lib/client/jsx/components/dashboard/dashboard.tsx
@@ -7,7 +7,7 @@ import {makeStyles} from '@material-ui/core/styles';
 
 import {Workflow} from '../../api_types';
 import WorkflowsCarousel from './workflows_carousel';
-import WorkspacesGrid from './workspace_control/workspaces_grid';
+import WorkspacesGrid from './workspaces_grid';
 import WorkspacesControls from './workspace_control/workspaces_controls';
 import WorkflowControls from './workflow_control/workflow_controls';
 import ProjectHeader from 'etna-js/components/project-header';
